requirements.txt
future~=0.18.2
chess~=1.9.4
numpy~=1.24.2
pillow~=9.2.0
keras~=2.9.0
requests~=2.27.1